+++ aws/eks-helm/aws-load-balancer-controller/iam.tf
@@ -22,7 +22,7 @@ resource "aws_iam_role" "aws_load_balancer_controller" {
 }
 
 resource "aws_iam_policy" "aws_load_balancer_controller" {
-  policy = file("${path.module}/policies/AWSLoadBalancerControllerIAMPolicy.json")
+  policy = file(local.policy_file)
   name   = "${var.cluster_name}-load-balancer-controller-policy"
 }