+++ aws/eks-helm/aws-load-balancer-controller/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [0.1.41]() (2025-01-08)
+# Features
+* Add new policy documents for chart versions and auto get policy version based on chart_version.
+
 ## [0.1.20]() (2024-12-24)
 ### Features
 * Initial commit with all the code